test: add assert statements to ensure index is greater than 0
Current methods in ParserStorage do not have any checks to ensure
whether the index of what they are getting is correct.

Any erroneous input from the user may cause the task to not be added
properly

Add assert statements to ensure that all the indexes are more than 0

This ensures that the data is properly parsed before being used.