refactor: add assertCheck method to command class and subclasses
Currently, command subclasses are not checking whether their
Storage, UI, userInput and taskList dependencies are being injected
properly before executing the commands in the method.

This needs to change as the user has the option of using two different
UI classes depending on what type of UI the user is using. If the
dependency is not injected properly, it will cause the application to
not work properly.

Add a new `testAssert(String userInput, TaskList taskList)` method to
the base command class and called it in all command subclasses.

This change improves the robustness of our commands without altering
their functional behavior.